reduce scope of ramp rate PR to only crosswalk analysis
